python-sidecar/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sentence-transformers==3.3.0
spacy==3.8.0
faiss-cpu==1.9.0
pydantic==2.10.0
numpy==1.26.4
